test(autocomplete): share the multiple selection fixture across suites
